Warnings fix for GCC15 and AOCC (Partial)
A subset of warnings have been fixed for GCC and AOCC

+++ BLACS/SRC/BI_BlacsWarn.c
@@ -20,6 +20,7 @@ void BI_BlacsWarn(Int ConTxt, Int line, char *file, char *form, ...)
    else myrow = mycol = -1;
 
    fprintf(stderr,
-"BLACS WARNING '%s'\nfrom {%d,%d}, pnum=%d, Contxt=%d, on line %d of file '%s'.\n\n",
-           cline, myrow, mycol, BI_Iam, ConTxt, line, file);
+"BLACS WARNING '%s'\nfrom {%ld,%ld}, pnum=%ld, Contxt=%ld, on line %ld of file '%s'.\n\n",
+           cline, (long)myrow, (long)mycol, (long)BI_Iam,
+           (long)ConTxt, (long)line, file);
 }
